+++ src/app/(fullscreen)/event-completed/page.tsx
@@ -1,0 +1,58 @@
+"use client";
+
+import Image from "next/image";
+import { useRouter } from "next/navigation";
+import { Button, FixedLayout } from "@/components";
+import { PATHS } from "@/constants";
+
+export default function EventCompletedPage() {
+  const router = useRouter();
+
+  const handleComplete = () => {
+    router.push(PATHS.FEEDBACK);
+  };
+
+  const handleCancel = () => {
+    router.push(PATHS.HOME);
+  };
+
+  return (
+    <>
+      <FixedLayout
+        title="이벤트 상영 완료"
+        showBackButton
+        isHeader
+        showBottomButton={false}
+        state="default"
+      >
+        <div className="mt-31 flex flex-col items-center justify-center text-center">
+          <div className="relative mb-6 aspect-square w-[200px]">
+            <Image
+              src="/images/image.png"
+              alt="이벤트 포스터"
+              fill
+              className="rounded-xl object-cover"
+            />
+          </div>
+          <div className="body-2-semibold text-white">
+            “더 폴: 오디언스와 환상의 문”
+          </div>
+          <div className="title-2-semibold mt-1 text-white">
+            이벤트 상영은 잘 완료되었나요?
+          </div>
+        </div>
+      </FixedLayout>
+
+      <div className="bg-gray-black fixed bottom-0 z-50 w-full max-w-125 px-5 pt-5 pb-12.5">
+        <div className="flex flex-col gap-3">
+          <Button className="bg-red-main text-white" onClick={handleComplete}>
+            네, 상영이 완료되었어요
+          </Button>
+          <Button className="bg-gray-800 text-white" onClick={handleCancel}>
+            아니요, 상영이 취소되었어요
+          </Button>
+        </div>
+      </div>
+    </>
+  );
+}
