+++ src/app/(fullscreen)/feedback/result/page.tsx
@@ -1,9 +1,10 @@
 import { Suspense } from "react";
 import FeedbackPage from "./components/feedbackpage";
+import Loading from "app/loading";
 
 export default function FeedbackResultPage() {
   return (
-    <Suspense fallback={<div>로딩 중...</div>}>
+    <Suspense fallback={<Loading />}>
       <FeedbackPage />
     </Suspense>
   );
